requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
cryptography==41.0.7
ecdsa==0.18.0
web3==6.11.3
python-dotenv==1.0.0
setuptools>=65.0.0
py-solc-x==2.0.4
